requirements.txt
openai>=1.0.0
rich>=13.0.0
backports.zoneinfo; python_version < "3.9"
